BAD MIGRATION CHANGEgit add .!

+++ db/migrate/20140809144601_add_cols_to_shoe.rb
@@ -3,7 +3,7 @@ class AddColsToShoe < ActiveRecord::Migration
     add_column :shoes, :brand, :string
     add_column :shoes, :description, :text
     add_column :shoes, :age_group, :string
-    add_column :shoes, :price, :decimal
+    add_column :shoes, :price, :float
     add_column :shoes, :product_type, :string
     add_column :shoes, :gender, :string
     add_column :shoes, :color, :string
